Resolved conflicts

+++ components/script/dom/performancepainttiming.rs
@@ -17,7 +17,7 @@ pub struct PerformancePaintTiming {
 }
 
 impl PerformancePaintTiming {
-    fn new_inherited(metric_type: ProgressiveWebMetricType, start_time: f64) -> PerformancePaintTiming {
+    fn new_inherited(metric_type: ProgressiveWebMetricType, start_time: u64) -> PerformancePaintTiming {
         let name = match metric_type {
             ProgressiveWebMetricType::FirstPaint => DOMString::from("first-paint"),
             ProgressiveWebMetricType::FirstContentfulPaint => DOMString::from("first-contentful-paint"),
@@ -26,7 +26,7 @@ impl PerformancePaintTiming {
         PerformancePaintTiming {
             entry: PerformanceEntry::new_inherited(name,
                                                    DOMString::from("paint"),
-                                                   start_time,
+                                                   start_time as f64,
                                                    0.)
         }
     }
@@ -34,7 +34,7 @@ impl PerformancePaintTiming {
     #[allow(unrooted_must_root)]
     pub fn new(global: &GlobalScope,
                metric_type: ProgressiveWebMetricType,
-               start_time: f64) -> DomRoot<PerformancePaintTiming> {
+               start_time: u64) -> DomRoot<PerformancePaintTiming> {
         let entry = PerformancePaintTiming::new_inherited(metric_type, start_time);
         reflect_dom_object(Box::new(entry), global, PerformancePaintTimingBinding::Wrap)
     }
